requirements.txt
click==6.7
pycrypto==2.6.1
requests==2.17.3
